+++ controller/product_controller.go
@@ -35,13 +35,14 @@ func (controller *ProductController) UpdateProduct(context *gin.Context) {
 	helper.ReadFromRequestBody(context, &productRequest)
 
 	productId, _ := strconv.Atoi(context.Param("id"))
-	product := controller.productService.Update(uint(productId), productRequest)
+	product := controller.productService.Update(productId, productRequest)
+
 	helper.JsonResponse(context, http.StatusOK, product)
 }
 
 func (controller *ProductController) DeleteProduct(context *gin.Context) {
 	productId, _ := strconv.Atoi(context.Param("id"))
-	err := controller.productService.Delete(uint(productId))
+	err := controller.productService.Delete(productId)
 	helper.PanicIfError(err)
 
 	helper.JsonResponse(context, http.StatusOK, struct {
